KAA-1423:  code review changes

+++ server/node/src/main/java/org/kaaproject/kaa/server/admin/services/ConfigurationServiceImpl.java
@@ -260,6 +260,7 @@ public class ConfigurationServiceImpl extends AbstractAdminService implements Co
 
     @Override
     public EndpointSpecificConfigurationDto deleteEndpointSpecificConfiguration(String endpointKeyHash) throws KaaAdminServiceException {
+        checkAuthority(KaaAuthorityDto.TENANT_DEVELOPER, KaaAuthorityDto.TENANT_USER);
         try {
             checkEndpointProfile(endpointKeyHash);
             EndpointSpecificConfigurationDto configuration = controlService.deleteEndpointSpecificConfiguration(endpointKeyHash);
